se definieron los metodos en el gerentePortIn

+++ src/main/java/com/byma/fondos_up_back/domain/model/Gerente.java
@@ -11,8 +11,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Gerente {
 
-    private Integer idRegistro;
-    private Integer idOrganizacionGerente;
+    private Long idRegistro;
+    private Long idOrganizacionGerente;
     private String denominacion;
     private boolean liquidaEnByma;
     private boolean habilitado;
