+++ src/main/java/com/byma/fondos_up_back/infrastructure/adapter/in/web/dto/request/GerenteRequestDTO.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class GerenteRequestDTO {
     @NotNull
-    private Integer idOrganizacionGerente;
+    private Long idOrganizacionGerente;
     @NotNull
     @NotBlank
     private String denominacion;
