+++ src/main/java/com/byma/fondos_up_back/infrastructure/adapter/in/web/dto/response/GerenteResponseDTO.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GerenteResponseDTO {
-    private Integer idRegistro;
-    private Integer idOrganizacionGerente;
+    private Long idRegistro;
+    private Long idOrganizacionGerente;
     private String denominacion;
     private Boolean liquidaEnByma;
     private Boolean habilitado;
